Hero Level, XP und LevelUp, Skills mit meheren Level. Skill Upgrade GUI. Zusammen gefasste Game Design-Datei, wo man Daten ändern kann für Lutz.
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1085

+++ inc/Game/Projectile.h
@@ -13,7 +13,7 @@ namespace TikiEngine
 		struct ProjectileDescription
 		{
 			/*! @brief how much damage it inflicts */
-			float Damage;
+			double Damage;
 
 			/*! @brief the target's position */
 			Vector3 Target;
@@ -37,7 +37,7 @@ namespace TikiEngine
 				Mass = 1;
 				MaxForce = 150;
 				Scale = 0.7f;
-				Damage = 5.0f;
+				Damage = 5.0;
                 LifeTime = 3.0f;
                 Shooter = 0;
 			}
@@ -74,7 +74,7 @@ namespace TikiEngine
 
 		protected:
 
-			float damage;
+			double damage;
 
 			Vector3 target;
 			Vector3 origin;
